•	Shortcoming: Incorect comment
•	Refactoring Method:
Remove Unused Code

•	Description:
The code includes comments that suggest the methods are unused, but they are actually used and serve a purpose. Therefore, the comments should be removed as they are misleading.

•	Why it is chosen:
Removing incorrect comments is essential for maintaining code accuracy and clarity. In this case, the methods are not unused, so the comments are misleading and can cause confusion. Removing them will improve code comprehension and prevent potential misunderstandings.

•	Conclusion:
By removing the incorrect comments, the code becomes more accurate and easier to understand. The absence of misleading comments enhances code readability by providing a clearer representation of the purpose and functionality of the methods.