•	Shortcoming: Dead Code
•	Refactoring Method:
Remove Dead Code

•	Description:
The main function is not used in the code and serves no purpose. It can be safely removed without affecting the functionality of the remaining code.

•	Why it is chosen:
Removing dead code is crucial for improving code maintainability and reducing unnecessary complexity. Unused code adds clutter and confusion, making the code harder to read and understand. By removing the unused main function, we can streamline the code and focus on the relevant parts.

•	Conclusion:
By removing the dead code (main function), the code becomes more concise and easier to comprehend. Removing the unnecessary code eliminates distractions and confusion caused by unused portions. The absence of dead code enhances code readability and improves overall code quality, making it easier for developers to understand and maintain the remaining relevantlogic