• Shortcoming: Primitive Obsession
• Refactoring Method:
Replace Primitive with Enum

• Description:
The refactoring method involved replacing the primitive string parameter with an enum called Position. The Position enum was created with all the possible positions in the game. The selectedPosition function was modified to accept a Position parameter instead of a string.

• Why it is chosen:
By using an enum, we provide clear constraints and type safety for the positions, eliminating the need for manual string comparisons. This approach improves code comprehension as the intent is more explicit, self-documenting, and less error-prone

• Conclusion:
The refactoring of replacing the primitive string parameter with an enum improves the readability of the code